fix(Notion): add APIErrorCode.ConflictError management

+++ test/__helpers__/teardown.ts
@@ -6,7 +6,8 @@ import { cleanTestTable } from './integrations/notion'
 async function globalTeardown() {
   await cleanTestTable()
   const isUnitTest = process.argv.includes('test/unit')
-  if (isUnitTest) return
+  const isIntegrationTest = process.argv.includes('test/integration')
+  if (isUnitTest || isIntegrationTest) return
   await container.stop()
   await fs.emptyDir(join(process.cwd(), 'tmp'))
 }
